+++ crates/rust-client/src/store/data_store.rs
@@ -81,7 +81,6 @@ impl DataStore for ClientDataStore {
                     err,
                 )
             })?;
-
         Ok((account, seed, block_header, partial_blockchain))
     }
 }