+++ crates/rust-client/src/tests.rs
@@ -497,15 +497,16 @@ async fn sync_state_tags() {
 
     // Import first mockchain note as expected
     let expected_notes = rpc_api.get_available_notes();
-
     for tag in expected_notes.iter().map(|n| n.metadata().tag()) {
         client.add_note_tag(tag).await.unwrap();
     }
 
-    // assert that we have no  expected notes prior to syncing state
+    // assert that we have no expected notes prior to syncing state
     assert!(client.get_input_notes(NoteFilter::Expected).await.unwrap().is_empty());
 
     // sync state
+    // The mockchain API has one public note and one private note, so in the end we will have
+    // the public one in the client
     let sync_details = client.sync_state().await.unwrap();
 
     // verify that the client is synced to the latest block
@@ -514,9 +515,9 @@ async fn sync_state_tags() {
         rpc_api.get_block_header_by_number(None, false).await.unwrap().0.block_num()
     );
 
-    // as we are syncing with tags, the response should contain blocks for both notes but they
-    // shouldn't be stored as they are not relevant for the client's accounts.
-    assert_eq!(client.test_store().get_tracked_block_headers().await.unwrap().len(), 0);
+    assert_eq!(client.get_input_notes(NoteFilter::All).await.unwrap().len(), 1);
+    // as we are syncing with tags, the response should contain blocks for both notes
+    assert_eq!(client.test_store().get_tracked_block_headers().await.unwrap().len(), 1);
 }
 
 #[tokio::test]
@@ -1742,7 +1743,7 @@ async fn swap_chain_test() {
 
     // Generate a few account pairs with a fungible asset that can be used for swaps.
     let mut account_pairs = vec![];
-    for _ in 0..5 {
+    for _ in 0..3 {
         let (wallet, faucet) =
             setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &keystore).await;
         mint_and_consume(&mut client, wallet.id(), faucet.id(), NoteType::Private).await;