+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## 0.10.0 (TBD)
 
-
 ### Changes
 
 * [BREAKING] Renamed `miden-cli` crate to `miden-client-cli`, and the `miden` executable to `miden-client` (#960).
@@ -14,6 +13,7 @@
 * [BREAKING] Renamed P2IDR to P2IDE (#1016).
 * Replaced deprecated #[clap(...)] with #[command(...)] and #[arg(...)] (#897).
 * [BREAKING] Removed `with_` prefix from builder functions (#1018).
+* Changed sync to store notes regardless of consumption checks if it matched a tracked tag (#1031).
 * Added a way to instantiate a `ScriptBuilder` from `Client` (#1022).
 
 ### Features